requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
coverage==6.4.1
dj-database-url==0.5.0
Django==4.0.4
django-bootstrap-v5==1.0.11
django-cors-headers==3.12.0
django-filter==21.1
django-on-heroku==1.1.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Faker==12.0.1
flake8==4.0.1
gunicorn==20.1.0
importlib-metadata==4.11.4
iniconfig==1.1.1
Markdown==3.3.7
mccabe==0.6.1
mixer==7.2.2
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.6
pytz
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
whitenoise==6.2.0
zipp==3.8.0
